+++ packages/charts/.storybook/main.ts
@@ -11,7 +11,7 @@ const getAbsolutePath = (packageName: string): string => {
 };
 
 const config: StorybookConfig = {
-    stories: ['../src/**/*.mdx', '../src/**/*.stories.@(ts|tsx)'],
+    stories: ['../src/**/*.stories.@(ts|tsx)'],
     staticDirs: ['assets'],
     addons: [
         '@etchteam/storybook-addon-status',