+++ storybook-docs/.storybook/main.ts
@@ -12,7 +12,7 @@ const getAbsolutePath = (packageName: string): string => {
 };
 
 const config: StorybookConfig = {
-    stories: ['../stories/**/*.mdx', '../stories/**/*.stories.@(ts|tsx)'],
+    stories: ['../stories/**/*.mdx'],
     addons: [
         '@storybook/addon-links',
         '@storybook/addon-a11y',