+++ packages/rte/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @frontify/fondue-rte
 
+## 0.0.1-rc.9
+
+### Patch Changes
+
+- [#2583](https://github.com/Frontify/fondue/pull/2583) [`0b953d7`](https://github.com/Frontify/fondue/commit/0b953d7e0098cefc60a845fd67e0b8d46fb5098c) Thanks [@noahwaldner](https://github.com/noahwaldner)! - feat: upgrade `@types/react`and adjust legacy `TooltipIcon` to work with new icons
+
+- Updated dependencies [[`8a33af1`](https://github.com/Frontify/fondue/commit/8a33af120b8379fdd5ee3164af9e8901ecf4f7cb), [`0b953d7`](https://github.com/Frontify/fondue/commit/0b953d7e0098cefc60a845fd67e0b8d46fb5098c), [`1ff2cab`](https://github.com/Frontify/fondue/commit/1ff2cab1dd80d104b762a78b978cfdc06b70da3b), [`888ae9b`](https://github.com/Frontify/fondue/commit/888ae9b0056762c802a9c60626b44cb71d7e925c), [`46016ad`](https://github.com/Frontify/fondue/commit/46016ad4ce7f7a05c5fc666311e5815eec791514), [`2f33abe`](https://github.com/Frontify/fondue/commit/2f33abe5b6b8a1c4719b5dca01519ac980292044), [`77e5ab9`](https://github.com/Frontify/fondue/commit/77e5ab9d24477a94865497f544d1ac06d221d4a1), [`e9f4ebd`](https://github.com/Frontify/fondue/commit/e9f4ebd640f775c5813bf77d4fe038bd18dd674f), [`cfc1142`](https://github.com/Frontify/fondue/commit/cfc11424a43ecb18547f4c6035c4dd56cc351799), [`493b333`](https://github.com/Frontify/fondue/commit/493b33342f695f43c3371a9d5ab852babb5ffca6)]:
+    - @frontify/fondue-components@25.0.0-rc.3
+    - @frontify/fondue-charts@7.0.0-rc.4
+    - @frontify/fondue-tokens@5.0.0-rc.3
+    - @frontify/fondue-icons@0.23.0-rc.2
+
 ## 0.0.1-rc.8
 
 ### Patch Changes